third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/picture-in-picture-event-listener-breakpoints-expected.txt

Tests Picture-in-Picture event listener breakpoints.


Running: testPlayVideoEventBreakpoint
Script execution paused.
Call stack:
    0) onVideoPlay (picture-in-picture-event-listener-breakpoints.js:24)
Event target: VIDEO
Script execution resumed.

Running: testEnterPictureInPictureEventBreakpoint
Script execution paused.
Call stack:
    0) onEnterPictureInPicture (picture-in-picture-event-listener-breakpoints.js:36)
Event target: VIDEO
Script execution resumed.

Running: testLeavePictureInPictureEventBreakpoint
Script execution paused.
Call stack:
    0) onLeavePictureInPicture (picture-in-picture-event-listener-breakpoints.js:48)
Event target: VIDEO
Script execution resumed.
